requirements.txt
redis==5.0.7
python-telegram-bot==13.7
python-dotenv==1.0.0
vk_api==11.9.9